fixed pose copy bug(Mat need deepcopy using copyTo method)

+++ include/common.h
@@ -25,6 +25,7 @@
 #define TEST
 // #define USE_ORB
 #define USE_SIFT
+
 #ifdef USE_SIFT
 #include <opencv2/xfeatures2d.hpp>
 #endif
@@ -40,7 +41,7 @@ using PnP = std::vector<std::pair<cv::Point2f,cv::Point3f>>;//2d_3d
 using AllKeyPoints = std::vector<std::vector<cv::KeyPoint>>;
 using AllDescriptors = std::vector<cv::Mat>;
 using AllMatches = std::vector<std::vector<cv::DMatch>>;
-using MapPoints = std::vector<cv::Point3d>;
+using MapPoints = std::vector<cv::Point3f>;
 
 // using IndexPoint3Pairs
